+++ src/lambdacd/new_ui.clj
@@ -5,7 +5,7 @@
             [hiccup.page :as page]
             [clojure.data.json :as json :only [write-str]]
             [lambdacd.presentation :as presentation]
-            [lambdacd.manualtrigger :as manualtrigger]
+            [lambdacd.steps.manualtrigger :as manualtrigger]
             [lambdacd.util :as util]
             [ring.util.response :as resp]
             [lambdacd.execution :as execution]