+++ src/lambdacd/server.clj
@@ -6,7 +6,7 @@
             [lambdacd.new-ui :as new-ui]
             [clojure.data.json :as json :only [write-str]]
             [lambdacd.presentation :as presentation]
-            [lambdacd.manualtrigger :as manualtrigger]
+            [lambdacd.steps.manualtrigger :as manualtrigger]
             [lambdacd.util :as util]
             [ring.util.response :as resp]
             [lambdacd.execution :as execution]))